pyproject.toml
[project]
name = "code-agents-knowledge-sharing"
version = "0.1.0"
description = "AI Agents Knowledge Sharing - Migrated to Agno 2.3.8 with Ollama"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno[ollama,openai]>=2.3.8",
    "openai>=1.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.3.0",
    "mcp>=0.1.0",
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "llama-index-core>=0.12.0",
    "llama-index-llms-openai>=0.6.0",
    "llama-index-llms-ollama>=0.5.0",
]

